Process_S3Data_and_NotificationAlerts/requirements.txt
aws-cdk-lib==2.50.0
constructs>=10.0.0,<11.0.0
nltk==3.6.7
